临时分支

+++ src/main/java/com/cn/app/chatgptbot/controller/blog/result/ArticleResult.java
@@ -4,7 +4,7 @@ public class ArticleResult {
     private Integer code;
     private String msg;
     private Object data;
-    private Integer total;
+    private Long total;
 
     public ArticleResult(Integer code, String msg, Object data) {
         this.code = code;
@@ -12,18 +12,18 @@ public class ArticleResult {
         this.data = data;
     }
 
-    public ArticleResult(Integer code, String msg, Object data, int total) {
+    public ArticleResult(Integer code, String msg, Object data, long total) {
         this.code = code;
         this.msg = msg;
         this.data = data;
         this.total = total;
     }
 
-    public Integer getTotal() {
+    public Long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(Long total) {
         this.total = total;
     }
